extract to pagetrimmer.ts and viewerhistory.ts using ES modules
