+++ viewer/viewer.html
@@ -42,7 +42,7 @@ See https://github.com/adobe-type-tools/cmap-resources
 <!-- This snippet is used in production (included from viewer.html) -->
 <link rel="resource" type="application/l10n" href="locale/locale.properties">
 <script src="/build/pdf.js" defer></script>
-<script src="/out/viewer/latexworkshop.js" defer></script>
+<script src="/out/viewer/latexworkshop.js" type="module"></script>
 <script src="viewer.js" defer></script>
   </head>
